invite auth route
