+++ src/extra/generateToken.ts
@@ -1,0 +1,9 @@
+import jwt from "jsonwebtoken";
+
+const generateToken = (id: string) => {
+  return jwt.sign({ id }, process.env.JWT_SECRET || "", {
+    expiresIn: "30d",
+  });
+};
+
+export default generateToken;
