+++ src/routes/inviteAuth.ts
@@ -1,0 +1,13 @@
+import { Router } from "express";
+
+import { auth } from "../middleware/authMiddleware";
+import { acceptInviteToProjectRegistration, inviteToProject } from "../controllers/auth/inviteAuth/inviteAuthController";
+import { isAdmin } from "../middleware/adminMiddleware";
+// const router = require("express").Router();
+const router = Router();
+
+//api/registration
+router.post("/project/:projectId/invite", auth, isAdmin, inviteToProject);
+
+router.post("/project/:projectId",acceptInviteToProjectRegistration )
+export default router;
