+++ src/routes/tickets.ts
@@ -6,6 +6,7 @@ import {
   updateTicket,
 } from "../controllers/ticketController";
 const router = Router();
+//api/ticket
 router.post("/:projectId/create-ticket", auth, createTicket);
 router.put("/:projectId/ticket/:ticketId", auth, updateTicket);
 router.delete("/:projectId/ticket/:ticketId", auth, deleteTicket);
