+++ src/server.ts
@@ -9,15 +9,16 @@ import dashboardRoute from "./routes/dashboard";
 import projectRoute from "./routes/projects";
 import testRoute from "./routes/testRoute";
 import ticketsRoute from "./routes/tickets";
+import inviteAuthRoute from "./routes/inviteAuth";
 import { connectDB } from "./config/db";
 
 connectDB();
 app.use(bodyParser.json());
 app.use("/api/user", authRoute);
 app.use("/api/dashboard", dashboardRoute);
-app.use("/api/projects", projectRoute);
+app.use("/api/project", projectRoute);
 app.use("/api/project", ticketsRoute);
-
+app.use("/api/registration", inviteAuthRoute);
 //testing route dev only
 app.use("/api/test/", testRoute);
